requirements.txt
openai
gpt_index
PyPDF2
gradio
transformers
pinecone-client